refactor(test): consolidate artifact output safety
