+++ tests/Unit/Artifact/ArtifactOutputSafetyTest.php
@@ -7,8 +7,10 @@ namespace Greenlight\Tests\Unit\Artifact;
 use Greenlight\Attribute\Test;
 use Greenlight\Config\ArtifactConfiguration;
 use Greenlight\Core\Artifact\AttachmentError;
+use Greenlight\Core\Artifact\StagedAttachment;
 use Greenlight\Core\Result\Outcome;
 use Greenlight\Core\Result\TestResult;
+use Greenlight\Core\Test\SkipTest;
 use Greenlight\Core\Test\TestId;
 use Greenlight\Expect\Expect;
 use Greenlight\Fixture\TempDirectory;
@@ -42,15 +44,7 @@ final readonly class ArtifactOutputSafetyTest
         $root = $this->tempDirectory->subdirectory('output-symlink');
         $outside = $this->tempDirectory->subdirectory('outside-output');
         \mkdir($root . '/published');
-        $store = ArtifactStore::open(
-            new ArtifactConfiguration($root . '/published'),
-            $root,
-            'run-1',
-        );
-        $id = new TestId('Example\EvidenceTest', 'publishesEvidence');
-        $attachments = $store->forAttempt($id, 1, new TestArtifactBudget());
-        $attachments->text('evidence.txt', 'body');
-        $staged = $attachments->seal()[0];
+        [$store, $id, $staged] = $this->stageEvidence($root, $root . '/published');
         $firstSegment = \explode('/', $staged->storageKey)[0];
         \mkdir($store->publicDirectory(), 0o777, true);
         \symlink($outside, $store->publicDirectory() . '/' . $firstSegment);
@@ -77,5 +71,95 @@ final readonly class ArtifactOutputSafetyTest
         } finally {
             $store->cleanup();
         }
+    }
+
+    #[Test]
+    public function nonDirectoryOutputEntriesBlockPublication(): void
+    {
+        $root = $this->tempDirectory->subdirectory('output-entry');
+        \mkdir($root . '/published');
+        [$store, $id, $staged] = $this->stageEvidence($root, $root . '/published');
+        $firstSegment = \explode('/', $staged->storageKey)[0];
+        \mkdir($store->publicDirectory(), 0o777, true);
+        $blocker = $store->publicDirectory() . '/' . $firstSegment;
+        \file_put_contents($blocker, 'keep');
+
+        try {
+            Expect::that(static fn(): TestResult => $store->publish(new TestResult(
+                $id,
+                Outcome::Failed,
+                0.1,
+                0,
+                attachments: [$staged],
+            )))
+                ->because('artifact publication requires directory-only output path segments')
+                ->toThrow(
+                    AttachmentError::class,
+                    message: 'Attachment output path contains a non-directory entry.',
+                );
+            Expect::that((string) \file_get_contents($blocker))
+                ->because('a rejected publication MUST NOT replace the blocking entry')
+                ->toBe('keep');
+            Expect::that(\is_file($store->session()->stagingDirectory . '/' . $staged->storageKey))
+                ->because('rejected evidence remains available for recovery')
+                ->toBeTrue();
+        } finally {
+            $store->cleanup();
+        }
+    }
+
+    #[Test]
+    public function anUnwritableOutputParentPreservesStagedEvidence(): void
+    {
+        $root = $this->tempDirectory->subdirectory('output-permission');
+        $readOnly = $root . '/read-only';
+        \mkdir($readOnly, 0o700);
+        \chmod($readOnly, 0o500);
+        \clearstatcache(true, $readOnly);
+
+        if (\is_writable($readOnly)) {
+            \chmod($readOnly, 0o700);
+            throw new SkipTest('The filesystem does not enforce directory write permissions.');
+        }
+
+        [$store, $id, $staged] = $this->stageEvidence($root, $readOnly . '/artifacts');
+
+        try {
+            Expect::that(static fn(): TestResult => $store->publish(new TestResult(
+                $id,
+                Outcome::Failed,
+                0.1,
+                0,
+                attachments: [$staged],
+            )))
+                ->because('an unwritable output parent MUST reject attachment publication')
+                ->toThrow(
+                    AttachmentError::class,
+                    matching: '/^Failed to create attachment output directory/',
+                );
+            Expect::that(\is_file($store->session()->stagingDirectory . '/' . $staged->storageKey))
+                ->because('rejected evidence MUST remain available for recovery')
+                ->toBeTrue();
+        } finally {
+            \chmod($readOnly, 0o700);
+            $store->cleanup();
+        }
+    }
+
+    /**
+     * @return array{ArtifactStore, TestId, StagedAttachment}
+     */
+    private function stageEvidence(string $workingDirectory, string $outputDirectory): array
+    {
+        $store = ArtifactStore::open(
+            new ArtifactConfiguration($outputDirectory),
+            $workingDirectory,
+            'run-1',
+        );
+        $id = new TestId('Example\EvidenceTest', 'publishesEvidence');
+        $attachments = $store->forAttempt($id, 1, new TestArtifactBudget());
+        $attachments->text('evidence.txt', 'body');
+
+        return [$store, $id, $attachments->seal()[0]];
     }
 }
